SH_HopDong/SH_HopDong/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Jinja2==3.1.4
SQLAlchemy==2.0.30
pydantic==2.7.3
pytesseract==0.3.10
Pillow==10.4.0
python-docx==1.1.2
PyMuPDF==1.24.7
orjson==3.10.3
pyodbc==4.0.39 